+++ src/components/Review/Review.module.scss
@@ -7,13 +7,10 @@
 
   width: 100%;
   max-width: var(--max-width);
-  padding: 0 10px;
 }
 
 @media (width >= 1024px) {
   .review {
     gap: 60px;
-
-    padding: 0 20px;
   }
 }
